requirements.txt
playwright==1.49.1
boto3>=1.34.0
requests>=2.32.0
python-dotenv>=1.0.0
